+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.3.1]
+### Improved
+- Added the possibility to specify the custom title of the delivery method for each country
+- Added some more Lithuanian translations
+
 ## [1.3.0]
 ### Improved
 - Reworked location retrieval method by first downloading locations to a temporary DB table and only after successful download, copying them to the main DB table